+++ example/quick_start.ml
@@ -7,7 +7,7 @@ let () =
     let application = "yourapp"
   end in
   let module M = Directories.Project_dirs (App_id) in
-  let option_value = function None -> "None" | Some v -> v in
+  let option_value = function None -> "None" | Some v -> Fpath.to_string v in
   Format.printf "cache dir  = `%s`@." (option_value M.cache_dir);
   Format.printf "config dir = `%s`@." (option_value M.config_dir);
   Format.printf "data dir   = `%s`@." (option_value M.data_dir)