+++ src/directories_common.ml
@@ -1,5 +1,3 @@
-let ( / ) = Filename.concat
-
 module type App_id = sig
   val qualifier : string
 
@@ -8,13 +6,7 @@ module type App_id = sig
   val application : string
 end
 
-(* TODO: remove once we drop 4.07 *)
-let option_map f = function None -> None | Some v -> Some (f v)
-
-(* TODO: remove once we drop 4.07 *)
-let option_bind o f = match o with None -> None | Some v -> f v
-
-let relative_opt dir = if Filename.is_relative dir then None else Some dir
+let relative_opt dir = if Fpath.is_rel dir then None else Some dir
 
 let getenv env =
   match Sys.getenv env with
@@ -22,7 +14,11 @@ let getenv env =
   | "" -> None
   | v -> Some v
 
-let getenvdir env = option_bind (getenv env) relative_opt
+let getenvdir env =
+  match getenv env with
+  | None -> None
+  | Some v -> (
+    match Fpath.of_string v with Error _ -> None | Ok v -> relative_opt v )
 
 let lower_and_replace_ws s replace =
   let s = String.trim s in