+++ src/macos/directories.ml
@@ -9,26 +9,33 @@ module Base_dirs () = struct
       match (Unix.getpwuid (Unix.getuid ())).Unix.pw_dir with
       | exception Unix.Unix_error _ -> None
       | exception Not_found -> None
-      | dir -> relative_opt dir )
-    | Some _dir as dir -> dir
+      | dir ->
+        let dir = Fpath.of_string dir |> Result.to_option in
+        Option.bind dir relative_opt )
+    | Some _ as dir -> dir
 
   (** $HOME/Library/Caches *)
-  let cache_dir = option_map (fun dir -> dir / "Library" / "Caches") home_dir
+  let cache_dir =
+    Option.map (fun dir -> Fpath.(dir / "Library" / "Caches")) home_dir
 
   (** $HOME/Library/Application Support *)
   let config_dir =
-    option_map (fun dir -> dir / "Library" / "Application Support") home_dir
+    Option.map
+      (fun dir -> Fpath.(dir / "Library" / "Application Support"))
+      home_dir
 
   (** $HOME/Library/Application Support *)
   let data_dir =
-    option_map (fun dir -> dir / "Library" / "Application Support") home_dir
+    Option.map
+      (fun dir -> Fpath.(dir / "Library" / "Application Support"))
+      home_dir
 
   (** $HOME/Library/Application Support *)
   let data_local_dir = data_dir
 
   (** $HOME/Library/Preferences *)
   let preference_dir =
-    option_map (fun dir -> dir / "Library" / "Preferences") home_dir
+    Option.map (fun dir -> Fpath.(dir / "Library" / "Preferences")) home_dir
 
   (** None *)
   let runtime_dir = None
@@ -46,7 +53,8 @@ module User_dirs () = struct
       user database) *)
   let home_dir = Base_dirs.home_dir
 
-  let concat_home_dir suffix = option_map (fun dir -> dir / suffix) home_dir
+  let concat_home_dir suffix =
+    Option.map (fun dir -> Fpath.(dir / suffix)) home_dir
 
   (** $HOME/Music *)
   let audio_dir = concat_home_dir "Music"
@@ -61,7 +69,9 @@ module User_dirs () = struct
   let download_dir = concat_home_dir "Downloads"
 
   (** $HOME/Library/Fonts *)
-  let font_dir = concat_home_dir ("Library" / "Fonts")
+  let font_dir =
+    let library_dir = concat_home_dir "Library" in
+    Option.map (fun dir -> Fpath.(dir / "Fonts")) library_dir
 
   (** $HOME/Pictures *)
   let picture_dir = concat_home_dir "Pictures"
@@ -90,7 +100,7 @@ module Project_dirs (App_id : App_id) = struct
   let project_path =
     Format.sprintf "%s.%s.%s" qualifier organization application
 
-  let concat_project_path = option_map (fun dir -> dir / project_path)
+  let concat_project_path = Option.map (fun dir -> Fpath.(dir / project_path))
 
   (** $HOME/Libary/Caches/<project_path> *)
   let cache_dir = concat_project_path Base_dirs.cache_dir